+++ integration-tester/tests/40-stop/stop.js
@@ -4,6 +4,8 @@ const infrastructure = require('../../config/infrastructure.js');
 const deployment = quilt.createDeployment();
 deployment.deploy(infrastructure);
 
-const containers = new quilt.Container('foo', 'google/pause')
-  .replicate(infrastructure.nWorker);
+const containers = [];
+for (let i = 0; i < infrastructure.nWorker; i += 1) {
+  containers.push(new quilt.Container('foo', 'google/pause'));
+}
 deployment.deploy(containers);