+++ integration-tester/tests/build-dockerfile/build-dockerfile.js
@@ -4,12 +4,14 @@ const infrastructure = require('../../config/infrastructure.js');
 const deployment = quilt.createDeployment();
 deployment.deploy(infrastructure);
 
-for (let i = 0; i < infrastructure.nWorker; i += 1) {
-  deployment.deploy(new quilt.Container('bar',
-    new quilt.Image(`test-custom-image${i}`,
-      'FROM alpine\n' +
-      `RUN echo ${i} > /dockerfile-id\n` +
-      'RUN echo $(cat /dev/urandom | tr -dc \'a-zA-Z0-9\' | ' +
-      'fold -w 32 | head -n 1) > /image-id'), {
-      command: ['tail', '-f', '/dev/null'] }).replicate(2));
+for (let workerIndex = 0; workerIndex < infrastructure.nWorker; workerIndex += 1) {
+  const image = new quilt.Image(`test-custom-image${workerIndex}`,
+    'FROM alpine\n' +
+    `RUN echo ${workerIndex} > /dockerfile-id\n` +
+    'RUN echo $(cat /dev/urandom | tr -dc \'a-zA-Z0-9\' | ' +
+    'fold -w 32 | head -n 1) > /image-id');
+  for (let containerIndex = 0; containerIndex < 2; containerIndex += 1) {
+    deployment.deploy(new quilt.Container(
+      'bar', image, { command: ['tail', '-f', '/dev/null'] }));
+  }
 }